change Int to Long for identidiers

+++ app/com/superxtra/notepad/model/User.scala
@@ -2,7 +2,7 @@ package com.superxtra.notepad.model
 
 import play.api.libs.json._
 
-case class User(id: Int, name: String, lastName: String, email: String)
+case class User(id: Long, name: String, lastName: String, email: String)
 
 object User {
   implicit val userJsonFormat: OFormat[User] = Json.format[User]
